Add basic layout and remove sample code
