+++ client/js/app/index.html
@@ -2,12 +2,16 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/src/app/assets/favicon.svg" />
+    <link
+      rel="icon"
+      type="image/svg+xml"
+      href="/src/app/assets/img/favicon.svg"
+    />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Vite App</title>
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/app/pages/main.jsx"></script>
+    <script type="module" src="/src/app/main.jsx"></script>
   </body>
 </html>
