+++ client/js/app/src/app/components/index.js
@@ -1,2 +1,2 @@
-export { App } from 'app/components/layout/app';
 export { Error } from 'app/components/layout/error';
+export { Layout } from 'app/components/layout/layout';
