+++ client/js/app/src/app/components/layout/layout.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { AppShell } from '@mantine/core';
+import { Header } from 'app/components/layout/header';
+
+export function Layout({ children }) {
+  return (
+    <AppShell header={<Header />} fixed>
+      {children}
+    </AppShell>
+  );
+}
