+++ client/js/app/src/app/pages/home/home.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export function Home() {
+  return <>home</>;
+}
