--- client/js/app/src/app/pages/main.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import { App } from 'app/components';
-import 'app/styles/index.css';
-import { AppRouter } from 'app/libs/app-router';
-import { QueryBuilder } from 'app/pages/querybuilder/query-builder';
-
-ReactDOM.createRoot(document.getElementById('root')).render(
-  <React.StrictMode>
-    <AppRouter>
-      <App path="/" />
-      <QueryBuilder path="querybuilder" />
-    </AppRouter>
-  </React.StrictMode>
-);
